another small optimisation
